Edits to desugar and desugar2
Prefer to use while loop definitons for these and docstrings that are
specifications.  Note that the modified desugar2 does introduce one
more temp variable for the final result than would be done with the
original definiton. I think this is actually clearer and we're not
worried about efficiency here, but it could be tweaked to avoid this.